src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/attribute/reference/SingleReferenceFloatAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.single.attribute.SingleAttributeParser;
import org.openqa.selenium.WebElement;

public class SingleReferenceFloatAttributeParser extends SingleAttributeParser {

    public SingleReferenceFloatAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Float.parseFloat(element.getAttribute(super.attribute));
    }

}
