src/test/java/com/github/brunomndantas/jscrapper/support/parser/single/attribute/reference/SingleReferenceFloatAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.single.attribute.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class SingleReferenceFloatAttributeParserTest {

    @Test
    public void constructorTest() {
        String attribute = "";
        SingleReferenceFloatAttributeParser parser = new SingleReferenceFloatAttributeParser(attribute);
        assertEquals(null, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
        SingleReferenceFloatAttributeParser parser = new SingleReferenceFloatAttributeParser("");

        assertEquals(2f, parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return "2";
            }
        }));
    }

}
